+++ js/data.js
@@ -1,0 +1,42 @@
+const getInputData = () => {
+  const MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+  const DESCRIPTIONS = [
+    'Интересный момент.',
+    'Время остановилось.',
+    'Что-то особенное.',
+    'Взгляд в будущее.',
+    'Мгновение счастья.',
+    'Просто так.',
+    'Необычный кадр.',
+    'Вдохновение внутри.',
+    'Моменты жизни.',
+    'Вся суть в деталях.',
+    'Что-то важное.',
+    'Обычный день, необычное фото.',
+    'Время и пространство.',
+    'Просто красиво.',
+    'Мгновение для памяти.',
+    'Неожиданное сочетание.',
+    'Взгляд со стороны.',
+    'Всё начинается здесь и сейчас.',
+    'Момент, который стоит запомнить.',
+    'Немного магии в обычном дне.',
+    'В каждом кадре есть история.',
+    'Просто и красиво.',
+    'Вдохновляющий момент.',
+    'Обычное — необычное.',
+    'Всё, что нужно — это взгляд.'];
+
+  const NAMES = ['Алексей', 'Мария', 'Дмитрий', 'Елена', 'Иван', 'Ольга', 'Сергей', 'Анастасия', 'Михаил', 'Наталья', 'Владимир'];
+
+  return {DESCRIPTIONS, MESSAGES, NAMES};
+};
+
+export {getInputData};
